pyproject.toml
[project]
name = "airflow-repo-template"
version = "0.1.0"
description = "Modern Airflow 3 development template with uv, ruff, and prek"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow==3.1.5",
    "apache-airflow-providers-standard>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "AIR",    # Airflow-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # allow assert in tests
    "ARG",    # allow unused arguments in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=dags --cov=plugins --cov-report=term-missing"

[tool.coverage.run]
source = ["dags", "plugins"]
omit = ["*/tests/*", "*/__pycache__/*"]
